pyproject.toml
[tool.poetry]
name = "aimiko"
version = "0.0.1"
description = "A multipurpose bot designed for the Trans Gayming server"
authors = ["Noelle <73260931+No767@users.noreply.github.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
discord-py = {extras = ["voice"], version = "^2.3.1"}
python-dotenv = "^1.0.0"
aiodns = "^3.0.0"
brotli = "^1.0.9"
orjson = "^3.9.2"
asyncpg = "^0.28.0"
asyncpg-trek = "^0.3.1"
uvloop = {markers = "sys_platform != \"win32\"", version = "^0.17.0"}
discord-ext-menus = {git = "https://github.com/Rapptz/discord-ext-menus", rev = "8686b5d1bbc1d3c862292eb436ab630d6e9c9b53"}
python-dateutil = "^2.8.2"
parsedatetime = "^2.6"
lru-dict = "^1.2.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
pyright = "^1.1.318"
watchfiles = "^0.19.0"
jishaku = "^2.5.1"
ruff = "^0.0.280"


[tool.poetry.group.tests.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
nox = "^2023.4.22"
nox-poetry = "^1.0.3"
dpytest = "^0.7.0"


[tool.poetry.group.docs.dependencies]
sphinx = "^7.1.1"
sphinx-autobuild = "^2021.3.14"
sphinx-copybutton = "^0.5.2"
furo = "^2023.7.26"

[tool.isort]
profile = 'black'

[tool.pyright]
include = ["Bot/**"]
exclude = ["**/__pycache__", "**/.mypy_cache", "**/.dmpypy.json"]
ignore = ["Docker"]
reportMissingImports = true
reportMissingTypeStubs = false

[tool.bandit]
skips = ["B311", "B101"]

[tool.ruff]
ignore = ["E501"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
